ml/requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
mediapipe>=0.10.9
PyYAML>=6.0.1
Pillow>=10.0.0
tqdm>=4.66.0
pytest>=7.4.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Optional training (FL3D)
pandas>=2.0.0
scikit-learn>=1.3.0
